validation refactoring
